add cost product

+++ modules/entity/transaction/transaction_detail.entity.go
@@ -6,8 +6,8 @@ type TransactionDetail struct {
 	*gorm.Model
 
 	TransactionId   uint
-	ProductId       uint `json:"ProductId" form:"ProductId" validate:"required"`
-	RatingProductId uint `json:"RatingProductId" form:"RatingProductId"`
-	Qty             uint `json:"Qty" form:"Qty" validate:"required"`
-	SubTotalPrice   uint `json:"SubTotalPrice" form:"SubTotalPrice" validate:"required"`
+	ProductId       uint    `json:"ProductId" form:"ProductId" validate:"required"`
+	RatingProductId uint    `json:"RatingProductId" form:"RatingProductId"`
+	Qty             uint    `json:"Qty" form:"Qty" validate:"required"`
+	SubTotalPrice   float64 `json:"SubTotalPrice" form:"SubTotalPrice" validate:"required"`
 }
